# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aplicación web para Comisión EMPRENDE (Backend)",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://backend-emprende.herokuapp.com/"
        }
    ],
    "paths": {
...
        "/api/v1/videoconferencia": {
            "get": {
                "tags": [
                    "Videoconferencias"
                ],
                "summary": "Mostrar información de videoconferencias",
                "operationId": "mostrarVideoconferencias",
                "responses": {
                    "200": {
                        "description": "Mostrar información."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/v1/videoconferencia/create": {
            "post": {
                "tags": [
                    "Videoconferencias"
                ],
                "summary": "Crear información de videoconferencias",
                "description": "Crear la información de videoconferencias",
                "operationId": "crearVideoconferencia",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "url"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "url": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/videoconferencia/{videoconferencia}": {
            "get": {
                "tags": [
                    "Videoconferencias"
                ],
                "summary": "Mostrar información específica de una videoconferencia",
                "operationId": "mostrarVideoconferencia",
                "parameters": [
                    {
                        "name": "videoconferencia",
                        "in": "path",
                        "description": "id de videoconferencia",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/videoconferencia/{videoconferencia}/update": {
            "post": {
                "tags": [
                    "Videoconferencias"
                ],
                "summary": "Actualizar información de videoconferencias",
                "description": "Actualizar la información de videoconferencias",
                "operationId": "actualizarVideoconferencia",
                "parameters": [
                    {
                        "name": "videoconferencia",
                        "in": "path",
                        "description": "id del emprendimiento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "url"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "url": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "name": "Authorization",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://backend-tesis.test/oauth/authorize",
                        "tokenUrl": "http://backend-tesis.test/oauth/token",
                        "refreshUrl": "http://backend-tesis.test/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": [
        {
            "oauth2_security_example": [
                "read",
                "write"
            ],
            "passport": []
        }
    ]
}